Move component files to components dir
