+++ src/components/BeaconRecords.tsx
@@ -5,7 +5,7 @@ import React, { FunctionComponent } from "react";
 
 import { IBeaconsGateway } from "../gateways/IBeaconsGateway";
 import PageContent from "../components/PageContent";
-import { BeaconsTable } from "../components/BeaconsTable";
+import { BeaconsTable } from "./BeaconsTable";
 import PageHeader from "../components/PageHeader";
 
 const useStyles = makeStyles((theme: Theme) =>
